feat : post logo
