+++ apps/frontend/src/components/feed/AsideLeft.tsx
@@ -1,45 +1,22 @@
 import { useAuthContext } from "../../hooks/useAuthContext";
 import Avatar from "react-avatar";
 import { color} from "../../lib/utils";
-import { useState, useEffect } from 'react'
 import ProfileLoder from "./ProfileLoder";
 import AuthLoder from "../auth/AuthLoder";
+import usefetchProfile from "../../hooks/usefetchProfile";
 
-import axios from 'axios';
 
 
 
 const AsideLeft = () => {
     const { state } = useAuthContext();
     const userId = state.user?.id || 'id';
-    const [imageUrl, setImageUrl] = useState<string | null>(null);
-    const [loading, setLoading] = useState<boolean>(false);
-    const [error, setError] = useState<string | null>(null);
-
-    const fetchProfile = async () => {
-        setLoading(true);
-        setError(null);
-        try {
-            const response = await axios.get(`${import.meta.env.VITE_API}/api/v0.1/user/${userId}`);
-            console.log(response.data);
-            setImageUrl(response.data.image);
-        } catch (error) {
-            console.error('Error fetching profile:', error);
-            setError('Failed to load profile');
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    useEffect(() => {
-        if (userId) {
-            fetchProfile();
-        }
-    }, [userId]);
+    const { imageUrl, loading, error } = usefetchProfile(userId);
 
     if (error) {
-        return <AuthLoder />
+        return <AuthLoder />;
     }
+
 
 
     return (
