+++ apps/frontend/src/components/feed/Postloder.tsx
@@ -1,0 +1,12 @@
+
+
+const Postloder = () => {
+    return (
+        <div className="">
+            <div className="h-12 w-12 lg:mx-auto rounded-full bg-slate-400"></div>
+
+        </div>
+    )
+}
+
+export default Postloder
